+++ src/components/plans/planComponents/PlanCreate.tsx
@@ -67,11 +67,6 @@ const PlanCreate = () => {
     setArea(e.target.value);
   };
 
-  // 제목
-  // const handleTitleChange = (e: any) => {
-  //   setTotalPlanTitle(e.target.value);
-  // };
-
   // 비용 변경 핸들러
   const handleBudgetChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const budgetValue = e.target.value.replace(/\D/g, ''); // 숫자 이외 삭제
