+++ src/components/reviews/reviewIndex/EditTrip.tsx
@@ -8,7 +8,7 @@ import {
   FieldContainer,
   Input,
   Label,
-} from '@/pages/travelReview/TravelReview.styles';
+} from '@/components/reviews/reviewIndex/TravelReviewPage.styles';
 
 const EditTrip = () => {
   const navigate = useNavigate();
