--- src/pages/travelPlan/TravelPlanMain.style.ts
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-const TravelPlanMainStyle = styled.div`
-  max-width: 1400px;
-  margin: 40px auto;
-  padding: 0px 20px;
-  margin-bottom: 500px;
-
-  h2 {
-    font-size: 24px;
-    padding-right: 50%;
-    font-weight: bold;
-  }
-`;
-
-export { TravelPlanMainStyle };
