+++ src/pages/travelPlan/TravelPlanUpdate1.tsx
@@ -1,4 +1,4 @@
-import * as S from './TravelPlanMain.style';
+import * as S from '../../components/plans/TravelPlanMain.style';
 import PlanUpdate from '@/components/plans/planComponents/PlanUpdate';
 
 const TravelPlanUpdate1 = () => {
