+++ src/pages/travelReview/ReviewCreate1.tsx
@@ -1,7 +1,7 @@
 // import ReDesignHeader from '@/components/layouts/Header2';
 import ReviewCreate from '@/components/reviews/reviewIndex/TReviewCreate';
 import { useState } from 'react';
-import styled from 'styled-components';
+import * as S from '@/components/reviews/reviewIndex/TReviewPageStyle';
 
 const ReviewCreate1 = () => {
   const [step, _] = useState<number>(1);
@@ -9,48 +9,15 @@ const ReviewCreate1 = () => {
   return (
     <>
       {/* <ReDesignHeader needSearchInput={true} /> */}
-      <CenteredContainer>
+      <S.CenteredContainer>
         <h2>여행 정보 작성하기</h2>
-        <StepperContainer>
-          <ProgressBar width={(step * 33).toString() + '%'} />
-        </StepperContainer>
+        <S.StepperContainer>
+          <S.ProgressBar width={(step * 33).toString() + '%'} />
+        </S.StepperContainer>
         <ReviewCreate />
-      </CenteredContainer>
+      </S.CenteredContainer>
     </>
   );
 };
 
 export default ReviewCreate1;
-
-const CenteredContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  width: 100%;
-  margin: 0 auto;
-  margin-top: 40px;
-  margin-bottom: 500px;
-
-  h2 {
-    font-size: 24px;
-    font-weight: bold;
-    margin-bottom: 40px;
-  }
-`;
-
-const StepperContainer = styled.div`
-  max-width: 700px;
-  width: 100%;
-  height: 20px;
-  background-color: #eee;
-  border-radius: 10px;
-  margin-bottom: 30px;
-`;
-
-const ProgressBar = styled.div<{ width: string }>`
-  border-radius: 10px;
-  height: 100%;
-  background-color: #5ac8ec;
-  width: ${({ width }) => width};
-`;
