+++ src/pages/travelReview/ReviewCreate3.tsx
@@ -1,58 +1,21 @@
-// import ReDesignHeader from '@/components/layouts/Header2';
+import * as S from '@/components/reviews/reviewIndex/TReviewPageStyle';
 import TReviewCreate3 from '@/components/reviews/reviewIndex/TReviewCreate3';
 import { useState } from 'react';
-// import React from 'react';
-import styled from 'styled-components';
 
 const ReviewCreate3 = () => {
   const [step, _] = useState<number>(3);
 
   return (
     <>
-      {/* <ReDesignHeader needSearchInput={true} /> */}
-      <CenteredContainer>
+      <S.CenteredContainer3>
         <h2>여행 정보 작성하기</h2>
-        <StepperContainer>
-          <ProgressBar width={(step * 33.5).toString() + '%'} />
-        </StepperContainer>
+        <S.StepperContainer>
+          <S.ProgressBar width={(step * 33.5).toString() + '%'} />
+        </S.StepperContainer>
         <TReviewCreate3 />
-      </CenteredContainer>
+      </S.CenteredContainer3>
     </>
   );
 };
 
 export default ReviewCreate3;
-
-const CenteredContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  width: 100%;
-  max-width: 1100px;
-  margin: 0 auto;
-  margin-top: 40px;
-  margin-bottom: 500px;
-
-  h2 {
-    font-size: 24px;
-    font-weight: bold;
-    margin-bottom: 40px;
-  }
-`;
-
-const StepperContainer = styled.div`
-  max-width: 700px;
-  width: 100%;
-  height: 20px;
-  background-color: #eee;
-  border-radius: 10px;
-  margin-bottom: 30px;
-`;
-
-const ProgressBar = styled.div<{ width: string }>`
-  border-radius: 10px;
-  height: 100%;
-  background-color: #5ac8ec;
-  width: ${({ width }) => width};
-`;
